feat: update user e address implementados

+++ src/services/users/deleteUser.service.ts
@@ -1,6 +1,7 @@
 import { Repository } from "typeorm";
 import { AppDataSource } from "../../data-source";
 import { User } from "../../entities";
+import { AppError } from "../../error";
 
 export const deleteUserServices = async (userId: number): Promise<void> => {
   const userRepository: Repository<User> = AppDataSource.getRepository(User);
@@ -8,6 +9,10 @@ export const deleteUserServices = async (userId: number): Promise<void> => {
   const user: User | null = await userRepository.findOneBy({
     id: userId,
   });
+
+  if (!user) {
+    throw new AppError("User not found", 404);
+  }
 
   await userRepository.remove(user!);
 };
